requirements.txt
torch
diffusers
transformers
Pillow
streamlit
